+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add handling for Rust crates:
   * Automatically download the packages referenced in a given `Cargo.lock` file.
   * Parse the metadata of `Cargo.toml` files.
+* Allow setting the log level for the CLI.
 
 # Version 0.9.0 - 2024-03-23